Added bears with abilities and example content
Check out the bears in the character playground map.
Switch between bears using 1, 2, 3, 4.
Press E to interact with interactables.
Press F to use the selected bear's special ability.

1: See Bear: Controls the camera.

2: Hear Bear: Press F to start listening to the heartbeat. Speed changes depending on the distance to BP_Heartbeat. Pressing F again or moving will stop the heartbeat.

3: Smell Bear: Press F to move to the closest BP_Special_Smell_Path. Once the bear is close, it will start to follow the path. Moving or pressing F again will stop following the path. Paths may not use any sharp turns, otherwise the bear will get stuck.

4: Throw Bear: Press F to grab an object or bear. Press F again to throw the object or bear. Cannot jump while carrying objects.

Every bear except for the see bear walks like a car. Not sure if / how I will fix this.
Design is not final. There is an example interactable and an example throwable included.